+++ lib/processor/sql.go
@@ -2,6 +2,7 @@ package processor
 
 import (
 	"database/sql"
+	"errors"
 	"fmt"
 	"sync"
 	"time"
@@ -10,11 +11,9 @@ import (
 	"github.com/Jeffail/benthos/v3/internal/bloblang/field"
 	"github.com/Jeffail/benthos/v3/internal/docs"
 	"github.com/Jeffail/benthos/v3/lib/log"
-	"github.com/Jeffail/benthos/v3/lib/message"
-	"github.com/Jeffail/benthos/v3/lib/message/tracing"
 	"github.com/Jeffail/benthos/v3/lib/metrics"
 	"github.com/Jeffail/benthos/v3/lib/types"
-	olog "github.com/opentracing/opentracing-go/log"
+	"github.com/opentracing/opentracing-go"
 
 	// SQL Drivers
 	_ "github.com/go-sql-driver/mysql"
@@ -29,8 +28,8 @@ func init() {
 			CategoryIntegration,
 		},
 		Summary: `
-Runs an SQL prepared query against a target database for each message batch and,
-for queries that return rows, replaces the batch with the result according to a
+Runs an SQL prepared query against a target database for each message and, for
+queries that return rows, replaces it with the result according to a
 [codec](#result-codecs).`,
 		Description: `
 If a query contains arguments they can be set as an array of strings supporting
@@ -41,8 +40,10 @@ the ` + "`args`" + ` field.
 
 The following is a list of supported drivers and their respective DSN formats:
 
-- ` + "`mysql`: `[username[:password]@][protocol[(address)]]/dbname[?param1=value1&...&paramN=valueN]`" + `
-- ` + "`postgres`: `postgresql://[user[:password]@][netloc][:port][/dbname][?param1=value1&...]`" + `
+| Driver | Data Source Name Format |
+|---|---|
+` + "| `mysql` | `[username[:password]@][protocol[(address)]]/dbname[?param1=value1&...&paramN=valueN]` |" + `
+` + "| `postgres` | `postgresql://[user[:password]@][netloc][:port][/dbname][?param1=value1&...]` |" + `
 
 Please note that the ` + "`postgres`" + ` driver enforces SSL by default, you
 can override this with the parameter ` + "`sslmode=disable`" + ` if required.`,
@@ -55,15 +56,14 @@ bar and baz populated with values extracted from messages:`,
 				Config: `
 pipeline:
   processors:
-    - for_each:
-      - sql:
-          driver: mysql
-          dsn: foouser:foopassword@tcp(localhost:3306)/foodb
-          query: "INSERT INTO footable (foo, bar, baz) VALUES (?, ?, ?);"
-          args:
-            - ${! json("document.foo") }
-            - ${! json("document.bar") }
-            - ${! meta("kafka_topic") }
+    - sql:
+        driver: mysql
+        data_source_name: foouser:foopassword@tcp(localhost:3306)/foodb
+        query: "INSERT INTO footable (foo, bar, baz) VALUES (?, ?, ?);"
+        args:
+          - ${! json("document.foo") }
+          - ${! json("document.bar") }
+          - ${! meta("kafka_topic") }
 `,
 			},
 			{
@@ -82,7 +82,7 @@ pipeline:
           - sql:
               driver: postgresql
               result_codec: json_array
-              dsn: postgres://foouser:foopass@localhost:5432/testdb?sslmode=disable
+              data_source_name: postgres://foouser:foopass@localhost:5432/testdb?sslmode=disable
               query: "SELECT * FROM footable WHERE user_id = $1;"
               args:
                 - ${! json("user.id") }
@@ -96,16 +96,18 @@ pipeline:
 				"A database [driver](#drivers) to use.",
 			).HasOptions("mysql", "postgres"),
 			docs.FieldCommon(
-				"dsn", "A Data Source Name to identify the target database.",
+				"data_source_name", "A Data Source Name to identify the target database.",
 				"foouser:foopassword@tcp(localhost:3306)/foodb",
+				"postgres://foouser:foopass@localhost:5432/foodb?sslmode=disable",
 			),
+			docs.FieldDeprecated("dsn"),
 			docs.FieldCommon(
 				"query", "The query to run against the database.",
 				"INSERT INTO footable (foo, bar, baz) VALUES (?, ?, ?);",
 			),
 			docs.FieldCommon(
 				"args",
-				"A list of arguments for the query to be resolved for each message batch.",
+				"A list of arguments for the query to be resolved for each message.",
 			).SupportsInterpolation(true),
 			docs.FieldCommon(
 				"result_codec",
@@ -116,12 +118,12 @@ pipeline:
 ## Result Codecs
 
 When a query returns rows they are serialised according to a chosen codec, and
-the batch contents are replaced with the serialised result.
+the message contents are replaced with the serialised result.
 
 ### ` + "`none`" + `
 
-The result of the query is ignored and the message batch remains unchanged. If
-your query does not return rows then this is the appropriate codec.
+The result of the query is ignored and the message remains unchanged. If your
+query does not return rows then this is the appropriate codec.
 
 ### ` + "`json_array`" + `
 
@@ -135,21 +137,23 @@ columns value in the row.`,
 
 // SQLConfig contains configuration fields for the SQL processor.
 type SQLConfig struct {
-	Driver      string   `json:"driver" yaml:"driver"`
-	DSN         string   `json:"dsn" yaml:"dsn"`
-	Query       string   `json:"query" yaml:"query"`
-	Args        []string `json:"args" yaml:"args"`
-	ResultCodec string   `json:"result_codec" yaml:"result_codec"`
+	Driver         string   `json:"driver" yaml:"driver"`
+	DataSourceName string   `json:"data_source_name" yaml:"data_source_name"`
+	DSN            string   `json:"dsn" yaml:"dsn"`
+	Query          string   `json:"query" yaml:"query"`
+	Args           []string `json:"args" yaml:"args"`
+	ResultCodec    string   `json:"result_codec" yaml:"result_codec"`
 }
 
 // NewSQLConfig returns a SQLConfig with default values.
 func NewSQLConfig() SQLConfig {
 	return SQLConfig{
-		Driver:      "mysql",
-		DSN:         "",
-		Query:       "",
-		Args:        []string{},
-		ResultCodec: "none",
+		Driver:         "mysql",
+		DataSourceName: "",
+		DSN:            "",
+		Query:          "",
+		Args:           []string{},
+		ResultCodec:    "none",
 	}
 }
 
@@ -165,6 +169,10 @@ type SQL struct {
 	dbMux    sync.Mutex
 	args     []field.Expression
 	resCodec sqlResultCodec
+
+	// TODO: V4 Remove this
+	deprecated         bool
+	resCodecDeprecated sqlResultCodecDeprecated
 
 	query *sql.Stmt
 
@@ -182,14 +190,22 @@ type SQL struct {
 func NewSQL(
 	conf Config, mgr types.Manager, log log.Modular, stats metrics.Type,
 ) (Type, error) {
-	resCodec, err := strToSQLResultCodec(conf.SQL.ResultCodec)
-	if err != nil {
-		return nil, err
+	deprecated := false
+	dsn := conf.SQL.DataSourceName
+	if len(conf.SQL.DSN) > 0 {
+		if len(dsn) > 0 {
+			return nil, errors.New("specified both a deprecated `dsn` as well as a `data_source_name`")
+		}
+		dsn = conf.SQL.DSN
+		deprecated = true
 	}
+
 	var db *sql.DB
-	if db, err = sql.Open(conf.SQL.Driver, conf.SQL.DSN); err != nil {
+	var err error
+	if db, err = sql.Open(conf.SQL.Driver, dsn); err != nil {
 		return nil, err
 	}
+
 	var args []field.Expression
 	for i, v := range conf.SQL.Args {
 		expr, err := bloblang.NewField(v)
@@ -198,13 +214,14 @@ func NewSQL(
 		}
 		args = append(args, expr)
 	}
+
 	s := &SQL{
 		log:        log,
 		stats:      stats,
 		conf:       conf.SQL,
 		db:         db,
 		args:       args,
-		resCodec:   resCodec,
+		deprecated: deprecated,
 		closeChan:  make(chan struct{}),
 		closedChan: make(chan struct{}),
 		mCount:     stats.GetCounter("count"),
@@ -212,10 +229,22 @@ func NewSQL(
 		mSent:      stats.GetCounter("sent"),
 		mBatchSent: stats.GetCounter("batch.sent"),
 	}
+
+	if deprecated {
+		if s.resCodecDeprecated, err = strToSQLResultCodecDeprecated(conf.SQL.ResultCodec); err != nil {
+			return nil, err
+		}
+	} else {
+		if s.resCodec, err = strToSQLResultCodec(conf.SQL.ResultCodec); err != nil {
+			return nil, err
+		}
+	}
+
 	if s.query, err = db.Prepare(conf.SQL.Query); err != nil {
 		db.Close()
 		return nil, fmt.Errorf("failed to prepare query: %v", err)
 	}
+
 	go func() {
 		defer func() {
 			s.db.Close()
@@ -231,9 +260,9 @@ func NewSQL(
 
 //------------------------------------------------------------------------------
 
-type sqlResultCodec func(rows *sql.Rows, msg types.Message) error
+type sqlResultCodec func(rows *sql.Rows, part types.Part) error
 
-func sqlResultJSONArrayCodec(rows *sql.Rows, msg types.Message) error {
+func sqlResultJSONArrayCodec(rows *sql.Rows, part types.Part) error {
 	columnNames, err := rows.Columns()
 	if err != nil {
 		return err
@@ -267,13 +296,7 @@ func sqlResultJSONArrayCodec(rows *sql.Rows, msg types.Message) error {
 		}
 		jArray = append(jArray, jObj)
 	}
-	if msg.Len() > 0 {
-		p := msg.Get(0)
-		msg.SetAll([]types.Part{p})
-		return msg.Get(0).SetJSON(jArray)
-	}
-	msg.Append(message.NewPart(nil))
-	return msg.Get(0).SetJSON(jArray)
+	return part.SetJSON(jArray)
 }
 
 func strToSQLResultCodec(codec string) (sqlResultCodec, error) {
@@ -299,51 +322,48 @@ func (s *SQL) doQuery(args ...interface{}) (*sql.Rows, error) {
 
 // ProcessMessage logs an event and returns the message unchanged.
 func (s *SQL) ProcessMessage(msg types.Message) ([]types.Message, types.Response) {
-	s.mCount.Incr(1)
-	result := msg.Copy()
-
-	spans := tracing.CreateChildSpans(TypeSQL, result)
-
-	args := make([]interface{}, len(s.args))
-	for i, v := range s.args {
-		args[i] = v.String(0, result)
+	if s.deprecated {
+		return s.processMessageDeprecated(msg)
 	}
-	var err error
-	if s.resCodec == nil {
-		if err = s.doExecute(args...); err != nil {
-			err = fmt.Errorf("failed to execute query: %v", err)
+
+	s.mCount.Incr(1)
+	newMsg := msg.Copy()
+
+	proc := func(index int, span opentracing.Span, part types.Part) error {
+		args := make([]interface{}, len(s.args))
+		for i, v := range s.args {
+			args[i] = v.String(index, msg)
 		}
-	} else {
-		var rows *sql.Rows
-		if rows, err = s.doQuery(args...); err == nil {
-			defer rows.Close()
-			if err = s.resCodec(rows, result); err != nil {
-				err = fmt.Errorf("failed to apply result codec: %v", err)
+		var err error
+		if s.resCodec == nil {
+			if err = s.doExecute(args...); err != nil {
+				err = fmt.Errorf("failed to execute query: %v", err)
 			}
 		} else {
-			err = fmt.Errorf("failed to execute query: %v", err)
+			var rows *sql.Rows
+			if rows, err = s.doQuery(args...); err == nil {
+				defer rows.Close()
+				if err = s.resCodec(rows, part); err != nil {
+					err = fmt.Errorf("failed to apply result codec: %v", err)
+				}
+			} else {
+				err = fmt.Errorf("failed to execute query: %v", err)
+			}
 		}
-	}
-	if err != nil {
-		result.Iter(func(i int, p types.Part) error {
-			FlagErr(p, err)
-			spans[i].LogFields(
-				olog.String("event", "error"),
-				olog.String("type", err.Error()),
-			)
-			return nil
-		})
-		s.log.Errorf("SQL error: %v\n", err)
-		s.mErr.Incr(1)
-	}
-	for _, s := range spans {
-		s.Finish()
+		if err != nil {
+			s.mErr.Incr(1)
+			s.log.Debugf("SQL error: %v\n", err)
+			return err
+		}
+		return nil
 	}
 
-	s.mSent.Incr(int64(result.Len()))
+	IteratePartsWithSpan(TypeSQL, nil, newMsg, proc)
+
 	s.mBatchSent.Incr(1)
-
-	return []types.Message{result}, nil
+	s.mSent.Incr(int64(newMsg.Len()))
+	msgs := [1]types.Message{newMsg}
+	return msgs[:], nil
 }
 
 // CloseAsync shuts down the processor and stops processing requests.